vimrc: ctrlp C-n binding remove <CR>
Without the <CR>, the C-n binding can be used for any use of CtrlP, for
example when you want to CtrlP for a file that's not in MRU or isn't in
the current dir, you can CtrlP to a specific dir. This also gives you
easier access to other CtrlP functions.