markdown.vim: control pmenu for ins C-n/p bindings
I've bound C-n/p in insert mode to up/down, but in markdown this changed
to up/down by visual line (gj and gk). When the pop-up menu is visible
(i.e., for completion), C-n/p would break out of the menu (as up/down)
instead of cycling through the menu. This fixes this behaviour.